% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaciesIdentificator-package.R
\docType{package}
\name{FaciesIdentificator-package}
\alias{FaciesIdentificator-package}
\title{FaciesIdentificator: Identifica facies sensu Cavallero et al. (2007) di rilievi vegetazionali}
\description{
Un pacchetto per l'identificazione delle facies da dati vegetazionali secondo il metodo descritto da Cavallero et al. (2007). Include un'applicazione Shiny per un'analisi interattiva.
}
\author{
\strong{Maintainer}: Marco Pittarello \email{marco.pittarello1987@gmail.com} (\href{https://orcid.org/0000-0001-6748-8790}{ORCID})

}
\keyword{internal}
